Refactor Yahoo Media RSS namespace patch
